[FIX] (stock_)delivery: move test for real invoice on back order
In this commit*, the forward-ported test was not moved with the fix on
stock.picking from delivery to stock_delivery, although the test relies
on real invoicing, which is defined in stock_delivery.

Fixes build error 22434

* e1e04f3a176e15fdce81ce3036093d3951401f52
