+++ addons/stock_delivery/tests/__init__.py
@@ -1,5 +1,6 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import test_carrier_propagation
+from . import test_delivery_cost
 from . import test_delivery_stock_move
 from . import test_packing_delivery
